Chrome requires WebTouchPoint to store WebFloatPoint, instead of WebPoint.
This is the third of three patches converting WebTouchPoint to use
WebFloatPoint instead of WebPoint. This cleans up some temporary
casts to WebPoint in chromium.

The first patch (https://codereview.chromium.org/140973005) allows
chromium to accept WebFloatPoint (it just casts all the floats to
ints).

The second patch was a blink modification to store WebFloatPoint
(https://codereview.chromium.org/149053002/).

BUG=336807

Review URL: https://codereview.chromium.org/148453012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258357 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9c769d419dc34ef2f7c146dd5cdbc7f81567cab2

+++ content/common/input/synthetic_web_input_event_builders.cc
@@ -159,7 +159,7 @@ void SyntheticWebTouchEvent::ResetPoints() {
   type = WebInputEvent::Undefined;
 }
 
-int SyntheticWebTouchEvent::PressPoint(int x, int y) {
+int SyntheticWebTouchEvent::PressPoint(float x, float y) {
   if (touchesLength == touchesLengthCap)
     return -1;
   WebTouchPoint& point = touches[touchesLength];
@@ -173,7 +173,7 @@ int SyntheticWebTouchEvent::PressPoint(int x, int y) {
   return point.id;
 }
 
-void SyntheticWebTouchEvent::MovePoint(int index, int x, int y) {
+void SyntheticWebTouchEvent::MovePoint(int index, float x, float y) {
   CHECK(index >= 0 && index < touchesLengthCap);
   WebTouchPoint& point = touches[index];
   point.position.x = point.screenPosition.x = x;
